requirements.txt
tensorflow==2.17.0
pillow
scikit-learn
matplotlib
